organizing the authentication helper functions so that there aren't multiple getAuthContext() functions, and the api.ts doesn't assume any particular context nesting in React.
